fix(java): use stable file locations for JARs provided by mason.nvim (#245)
Since the introduction of the new registry, packages started linking
more than executables (`bin/`) in Mason's root directory. Examples of
these are platform-agnostic "share" files, for example the JARs used
here.

Reaching into the internal package structures of Mason packages is not
stable and is prone to break across versions.