feat(nvim-devdocs): bundled `js/node` docs for typescript files (#834)
Previously, executing `:DevdocsOpenCurrentFloat` in a typescript file would only provide the Typescript docs. This is desirble, but most of the docs TS devs are interested in are normally under `javascript` (MDN) or `node`. You could work around this by opening up all docs, but if you use multiple languages docs, this could be a cumbersome experience.

This change bundles all three sets of docs to be avaiable under current `ts` floats, which should be a sane default for 99% of devs.