fix(catppuccin): set name to "catppuccin"
i tried installing hop as astrocommunity package, and in the repo i saw that it had code that supposedly should automatically enable integration for hop in catppuccin theme. yet that didnt work for some reason, and so i just enabled it manually. and today after thinking about it and 10 minutes of experimenting i figured it out. to fix this, all you have to do is set `name = "catppuccin"`, then its going to work. i just took hop as an example, but its with every single astrocommunity package i tried. like mini-clue or trouble and so on.